+++ examples/official-storybook/stories/addon-docs/forward-ref.stories.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { DocgenButton } from '../../components/DocgenButton';
 
-const ForwardedButton = React.forwardRef((props, ref) => <DocgenButton ref={ref} {...props} />);
+const ForwardedButton = React.forwardRef((props = { label: '' }, ref) => (
+  <DocgenButton ref={ref} {...props} />
+));
 
 export default {
   title: 'Addons|Docs/ForwardRef',
@@ -9,5 +11,5 @@ export default {
   parameters: { chromatic: { disable: true } },
 };
 
-export const displaysCorrectly = () => <ForwardedButton>Hello World!</ForwardedButton>;
+export const displaysCorrectly = () => <ForwardedButton label="hello" />;
 displaysCorrectly.story = { name: 'Displays forwarded ref components correctly' };