+++ lib/components/src/blocks/PropsTable/PropDef.ts
@@ -23,6 +23,7 @@ export type PropDefaultValue = PropSummaryValue;
 export interface PropDef {
   name: string;
   type: PropType;
+  sbType?: any;
   required: boolean;
   description?: string;
   defaultValue?: PropDefaultValue;